Adding necessary files to create documentation
Added GitHub Actions workflow file along with the
necessary files to create a documentation page with
Sphinx. Currently only using the default alabaster
theme, which works for now.

Only adding documentation for work on linear regression
analysis on the Ames, Iowa housing data set. Still have
to do some more work on the documentation on the origin
of the data set.