Making changes to start designing a GUI
Made a few modifications to the pre-processing notebook
to add a few columns that are based on the values of
other columns and will be used in the GUI. Some of these
values are more categorical/boolean to make it a simple
yes or no if the thing exists, like a garage.

This new change also requires creating a new ML model
using a subset of the columns that are being hand-picked
to decrease the amount of duplicated columns and also so
that the values being entered by a user make sense and
are not so technical, such as square footage of a garage.
A lot of this is inspired by what I am seeing and making
an honest attempt to make a tool that an every day person
can end up using if it ever becomes relevant.